Improve dialog.open error handling.
Slack can return a 200 OK to requests that do not succeed.
They return a JSON body with keys "ok", and in the case of
errors also a key "error".

This commit adds logging of those errors to Cloud Function logs.
It also addresses a common case where interactivity/dialogs have
not been enabled for the Slack App, and states that in the logs.